fix warnings

+++ src/aimlab/gameloop.hpp
@@ -69,7 +69,7 @@ public:
             targetsToSpawn.clear();
 
             // 객체 삭제 Swap And Pop 방식
-            for (int i = world3d.size() - 1; i > -1; --i)
+            for (size_t i = world3d.size() - 1; i > -1; --i)
             {
                 if (world3d[i]->state == Died)
                 {
